# perf(start): improve typescript performance of start types