This reduces instantiations of start types and some router types. Mostly middleware/server fns should type check faster after caching but also sprinkled in some variance annotations to help the language service for middleware

Before
```> tsc --extendedDiagnostics

Files:                        1175
Lines of Library:            40097
Lines of Definitions:       198118
Lines of TypeScript:         32262
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Identifiers:                314481
Symbols:                    508239
Types:                      156072
Instantiations:            2714633
Memory used:               587444K
Assignability cache size:    67144
Identity cache size:         15480
Subtype cache size:          44949
Strict subtype cache size:      10
I/O Read time:               0.16s
Parse time:                  0.96s
ResolveModule time:          0.54s
ResolveTypeReference time:   0.02s
ResolveLibrary time:         0.02s
Program time:                1.93s
Bind time:                   0.42s
Check time:                  7.66s
printTime time:              0.00s
Emit time:                   0.00s
Total time:                 10.00s
```

After
```
> tsc --extendedDiagnostics

Files:                        1175
Lines of Library:            40097
Lines of Definitions:       198133
Lines of TypeScript:         32262
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Identifiers:                314567
Symbols:                    502683
Types:                      148047
Instantiations:            2145246
Memory used:               572880K
Assignability cache size:    60557
Identity cache size:         15475
Subtype cache size:          44949
Strict subtype cache size:       9
I/O Read time:               0.16s
Parse time:                  0.91s
ResolveModule time:          0.54s
ResolveTypeReference time:   0.02s
ResolveLibrary time:         0.02s
Program time:                1.96s
Bind time:                   0.41s
Check time:                  6.91s
printTime time:              0.00s
Emit time:                   0.00s
Total time:                  9.27s
```